# migrate to cargo-metadata