--- a/src/racer/fileres.rs
+++ b/src/racer/fileres.rs
@@ -5,18 +5,10 @@ use std::path::{Path, PathBuf};
 /// get crate file from current path & crate name
 pub fn get_crate_file(name: &str, from_path: &Path, session: &Session) -> Option<PathBuf> {
     debug!("get_crate_file {}, {:?}", name, from_path);
-
-    if let Some(path) = get_outer_crates(name, from_path, session) {
-        debug!("get_outer_crates returned {:?} for {}", path, name);
-        return Some(path);
-    } else {
-        debug!("get_outer_crates returned None, try RUST_SRC_PATH");
-    }
-    get_std_file(name, session)
+    get_std_file(name, session).or_else(|| get_outer_crates(name, from_path, session))
 }
 
 pub fn get_std_file(name: &str, session: &Session) -> Option<PathBuf> {
-    // TODO: cache std libs
     if let Some(ref std_path) = *RUST_SRC_PATH {
         // try lib<name>/lib.rs, like in the rust source dir
         let cratelibname = format!("lib{}", name);
@@ -24,14 +16,8 @@ pub fn get_std_file(name: &str, session: &Session) -> Option<PathBuf> {
         if filepath.exists() || session.contains_file(&filepath) {
             return Some(filepath);
         }
-
-        // try <name>/lib.rs
-        let filepath = std_path.join(name).join("lib.rs");
-        if filepath.exists() || session.contains_file(&filepath) {
-            return Some(filepath);
-        }
     }
-    None
+    return None;
 }
 
 /// get module file from current path & crate name
@@ -59,5 +45,7 @@ fn get_outer_crates(libname: &str, from_path: &Path, session: &Session) -> Optio
     );
 
     let manifest = session.project_model.discover_project_manifest(from_path)?;
-    session.project_model.resolve_dependency(&manifest, libname)
+    let res = session.project_model.resolve_dependency(&manifest, libname);
+    println!("{:?}", res);
+    res
 }