--- a/test_project/test-crate3/src/main.rs
+++ b/test_project/test-crate3/src/main.rs
@@ -1,5 +1,4 @@
 extern crate rand;
-extern crate rayon;
 
 fn main() {
     println!("Hello, world!");