To remove cargo dependency.