@kngwyu 2018-08-26
benchmark result
before:
```
test completes_rand  ... bench:  99,061,180 ns/iter (+/- 5,382,882)
test completes_rayon ... bench:  14,012,226 ns/iter (+/- 2,249,149)
```
after: 
```
test completes_rand  ... bench:  66,768,857 ns/iter (+/- 3,299,046)
test completes_rayon ... bench:  29,467,723 ns/iter (+/- 3,618,528)
```

Since I couldn't see a serious performance issue here, I decided to merge this PR.