@birkenfeld 2018-08-26
A 100% worse number on `rayon` is not serious?