@kngwyu 2018-08-26
@birkenfeld 
Now we have a lot of issues around installing because of cargo(#927, #919), and it's more serious.
Indeed using subprocess make the program slower, but I think 0.01s delay is acceptable.
In addition, since we call `cargo metadata` only once in a session, it shouldn't be a large bottleneck.