@birkenfeld 2018-08-26
Well, for many users each completion is a new session. But as long as it's a constant offset that should be ok, at least as long as it doesn't dominate the processing time.